Moved out spotify routes to routes/spotify.js
