+++ index.js
@@ -5,7 +5,7 @@ const { google } = require('googleapis');
 const path = require('path');
 const config = require('./config');
 const { youtube } = require('googleapis/build/src/apis/youtube');
-const { generateCodeVerifier, generateCodeChallenge, getToken, fetchProfile, fetchPlaylists } = require('./lib/spotify_api');
+const spotify_routes = require('./routes/spotify');
 const { yt_request_all_playlists, isTokenValid } = require('./lib/yt_api');
 
 const app = express();
@@ -45,55 +45,7 @@ app.get("/api", (req, res) => {
   res.send(status);
 });
 
-app.get("/api/spotify/auth", async (req, res) => {
-  const verifier = generateCodeVerifier(128);
-  const challenge = await generateCodeChallenge(verifier);
-
-  req.session.user = { verifier: verifier }
-
-  const params = new URLSearchParams();
-
-  params.append("response_type", "code");
-  params.append("client_id", config.spotify.client_id);
-  params.append("scope", "user-read-private user-read-email");
-  params.append("code_challenge", challenge);
-  params.append("code_challenge_method", "S256");
-  params.append("redirect_uri", `${config.server_address}/api/spotify/callback`);
-
-  res.redirect(`https://accounts.spotify.com/authorize?${params.toString()}`);
-});
-
-app.get("/api/spotify/callback", async (req, res) => {
-  var code = req.query.code;
-  var verifier = req.session.user.verifier;
-
-  const accessToken = await getToken(verifier, code);
-
-  req.session.user = { access_token: accessToken };
-
-  const status = {
-    "status": "success",
-    "access_token": accessToken
-  };
-
-  res.send(status);
-});
-
-app.get("/api/spotify/list", async (req, res) => {
-  var accessToken = req.session.user.access_token;
-
-  var profile = await fetchProfile(accessToken);
-  var user_id = profile.id;
-
-  var playlists = await fetchPlaylists(user_id, accessToken);
-
-  const status = {
-    "status": "success",
-    "playlists": playlists
-  };
-
-  res.send(status);
-});
+app.use('/api/spotify', spotify_routes);
 
 app.get("/api/ytmusic/auth", async (req, res) => {
   var OAuth2 = google.auth.OAuth2;
